Sort out normalisation stuff
We've had some issues with the normalisation step when creating the term-document matrix. In particular, the normalisation as part of the TermDocumentMatrix command in the tm package doesn't really work. When we take it out, or replace it with a correct normalisation, we get a weird effect: a question by Barry Sheerman about prison innumeracy is rated as being waaaaay more important to the word 'prison' than any other question, because when the dimensionality reduction happens the values for innumeracy are squashed.

I haven't been able to work out why this occurs.

Absent a correct solution, I suggest that instead we reduce the dimensionality of the LSA space to 2000 rather than 1000, and do the normalisation correctly. Note that we keep the number of clusterrs at 1000, so this means having two parameters, one for LSA dimensionality and one for number of clusters.

This code does that, which means changes to:

 - README file to say the new process for defining the LSA dimensionality separate to the number of clusters.
 - DataCreator file
 - obviously the data changes as a result of the above