Data updated on 2/1/18 (#201)
There are some mismatching questions: one is not in our archive yet (was added online while the search space was being created), the other has some html class tags in the answer text online that aren't in our archive. I am not especially worried about either, but will repull the whole archive later anyway. But for now this will do.