Added the end-of-word tag `\\b` to line 125 in `Functions.R`
This line was intended to swap out occurrences of "disabl" that hadn't been unstemmed, which occurred where the original word was "disability" which stems to "disabl" but doesn't unstem. As it was this line had the unintended consequence of turning the correctly-unstemmed "disabled" into "disabilityed". Hopefully now it should work.