Data update 9/2/18: 8,961 questions (#227)
* Small changes to `global.R` and `DataCreator.R`:

 - removed pointless loading of antique data that isn't used in `global.R`
 - added some spaces around an "=" in `DataCreator.R`

* Added the end-of-word tag `\\b` to line 125 in `Functions.R`

This line was intended to swap out occurrences of "disabl" that hadn't been unstemmed, which occurred where the original word was "disability" which stems to "disabl" but doesn't unstem. As it was this line had the unintended consequence of turning the correctly-unstemmed "disabled" into "disabilityed". Hopefully now it should work.

* Data Update 9/2/18: 8,961 questions